+++ chains/substrate/pallet/pallet.go
@@ -1,7 +1,7 @@
 // The Licensed Work is (c) 2022 Sygma
 // SPDX-License-Identifier: BUSL-1.1
 
-package calls
+package pallet
 
 import (
 	"math/big"
